pyproject.toml
[project]
name = "h2h-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.89",
    "fastapi[standard]>=0.116.1",
    "firebase-admin>=6.0.0",
    "mangum>=0.17.0",
    "motor>=3.7.1",
    "pydantic>=2.12.0",
    "python-dotenv>=0.9.9",
]
